.CPP substituído por .C Registradores reservados salvos em strings
