+++ .gitignore
@@ -1,0 +1,4 @@
+
+*.exe
+
+vocabulario.exe
