+++ registradoresreservados.txt
@@ -1,0 +1,256 @@
+W
+F
+INDF
+TMR0
+PCL
+STATUS
+FSR
+PORTA
+PORTB
+PCLATH
+INTCON
+PIR1
+PIR2
+TMR1
+TMR1L
+TMR1H
+T1CON
+TMR2
+T2CON
+SSPBUF
+SSPCON
+CCPR1
+CCPR1L
+CCPR1H
+CCP1CON
+ADRESH
+ADCON0
+OPTION_REG
+TRISA
+TRISB
+PIE1
+PIE2
+PCON
+OSCCON
+OSCTUNE
+PR2
+SSPADD
+SSPSTAT
+ADRESL
+ADCON1
+EEDATA
+EEADR
+EEDATH
+EEADRH
+EECON1
+EECON2
+C
+DC
+Z
+NOT_PD
+NOT_TO
+IRP
+RP0
+RP1
+RA0
+RA1
+RA2
+RA3
+RA4
+RA5
+RA6
+RA7
+RB0
+RB1
+RB2
+RB3
+RB4
+RB5
+RB6
+RB7
+RBIF
+INTF
+TMR0IF
+RBIE
+INTE
+TMR0IE
+PEIE
+GIE
+TMR1IF
+TMR2IF
+CCP1IF
+SSPIF
+ADIF
+EEIF
+TMR1ON
+TMR1CS
+NOT_T1SYNC
+T1OSCEN
+T1CKPS0
+T1CKPS1
+T1INSYNC
+TMR2ON
+T2CKPS0
+T2CKPS1
+TOUTPS0
+TOUTPS1
+TOUTPS2
+TOUTPS3
+CKP
+SSPEN
+SSPOV
+WCOL
+SSPM0
+SSPM1
+SSPM2
+SSPM3
+CCP1Y
+CCP1X
+CCP1M0
+CCP1M1
+CCP1M2
+CCP1M3
+ADON
+GO_NOT_DONE
+GO
+CHS0
+CHS1
+CHS2
+ADCS0
+ADCS1
+NOT_DONE
+GO_DONE
+PSA
+T0SE
+T0CS
+INTEDG
+NOT_RBPU
+PS0
+PS1
+PS2
+TRISA0
+TRISA1
+TRISA2
+TRISA3
+TRISA4
+TRISA5
+TRISA6
+TRISA7
+TRISB0
+TRISB1
+TRISB2
+TRISB3
+TRISB4
+TRISB5
+TRISB6
+TRISB7
+TMR1IE
+TMR2IE
+CCP1IE
+SSPIE
+ADIE
+EEIE
+NOT_BOR
+NOT_POR
+NOT_BO
+IOFS
+IRCF0
+IRCF1
+IRCF2
+TUN0
+TUN1
+TUN2
+TUN3
+TUN4
+TUN5
+BF
+UA
+R_NOT_W
+S
+P
+D_NOT_A
+CKE
+SMP
+R
+D
+I2C_READ
+I2C_START
+I2C_STOP
+I2C_DATA
+NOT_W
+NOT_A
+NOT_WRITE
+NOT_ADDRESS
+R_W
+D_A
+READ_WRITE
+DATA_ADDRESS
+ADCS2
+ADFM
+PCFG0
+PCFG1
+PCFG2
+PCFG3
+RD
+WR
+WREN
+WRERR
+FREE
+EEPGD
+_CONFIG
+_FOSC_LP
+_LP_OSC
+_FOSC_XT
+_XT_OSC
+_FOSC_HS
+_HS_OSC
+_FOSC_EC
+_EXTCLK
+_FOSC_INTOSCIO
+_INTRC_IO
+_FOSC_INTOSCCLK
+_INTRC_CLKOUT
+_FOSC_EXTRCIO
+_EXTRC_IO
+_FOSC_EXTRCCLK
+_EXTRC_CLKOUT
+_WDTE_OFF
+_WDT_OFF
+_WDTE_ON
+_WDT_ON
+_PWRTE_ON
+_PWRTE_OFF
+_MCLRE_OFF
+_MCLR_OFF
+_MCLRE_ON
+_MCLR_ON
+_BOREN_OFF
+_BODEN_OFF
+_BOREN_ON
+_BODEN_ON
+_LVP_OFF
+_LVP_ON
+_CPD_ON
+_CPD_OFF
+_WRT_1536
+_WRT_ENABLE_1536
+_WRT_1024
+_WRT_ENABLE_1024
+_WRT_512
+_WRT_ENABLE_512
+_WRT_OFF
+_WRT_ENABLE_OFF
+_DEBUG_ON
+_DEBUG_OFF
+_CCPMX_RB3
+_CCP1_RB3
+_CCPMX_RB2
+_CCP1_RB2
+_CP_ON
+_CP_ALL
+_CP_OFF
+_DEVID1
+_IDLOC0
+_IDLOC1
+_IDLOC2
+_IDLOC3
